+++ packages/contracts/contracts/core/IDAO.sol
@@ -16,7 +16,7 @@ abstract contract IDAO {
     /// @param _where The address of the contract.
     /// @param _who The address of a EOA or contract to give the permissions.
     /// @param _permissionId The permission identifier.
-    /// @param _data The optional data passed to the `PermissionOracle` registered.
+    /// @param _data The optional data passed to the `PermissionCondition` registered.
     /// @return bool Returns true if the address has permission, false if not.
     function hasPermission(
         address _where,
@@ -41,7 +41,7 @@ abstract contract IDAO {
     /// @return bytes[] The array of results obtained from the executed actions in `bytes`.
     /// @return uint256 The constructed failureMap which contains which actions have actually failed.
     function execute(
-        uint256 callId,
+        bytes32 callId,
         Action[] memory _actions,
         uint256 _allowFailureMap
     ) external virtual returns (bytes[] memory, uint256);
@@ -56,7 +56,7 @@ abstract contract IDAO {
     /// @param execResults Array with the results of the executed actions.
     event Executed(
         address indexed actor,
-        uint256 callId,
+        bytes32 callId,
         Action[] actions,
         uint256 failureMap,
         bytes[] execResults