+++ packages/contracts/test/factory/dao-factory.ts
@@ -17,7 +17,7 @@ import {
 } from '../test-utils/repo';
 import {findEvent} from '../../utils/event';
 import {getMergedABI} from '../../utils/abi';
-import {deployNewDAO} from '../test-utils/dao';
+import {daoExampleURI, deployNewDAO} from '../test-utils/dao';
 import {deployWithProxy} from '../test-utils/proxy';
 
 const EVENTS = {
@@ -211,6 +211,7 @@ describe('DAOFactory: ', function () {
       trustedForwarder: AddressZero,
       name: daoDummyName,
       metadata: daoDummyMetadata,
+      daoURI: daoExampleURI,
     };
 
     votingSettings = {